requirements.txt
flask
python-dotenv
google-genai
faiss-cpu
pypdf
numpy
regex